# tools: Add watch mode to webpack-make