--- a/tools/webpack-make
+++ b/tools/webpack-make
@@ -22,13 +22,9 @@ try {
 }
 
 var ops = stdio.getopt({
+    config: { key: "c", args: 1, description: "Path to webpack.config.js", default: "webpack.config.js" },
     deps: { key: "d", args: 1, description: "Output dependencies in Makefile format" }
 });
-
-if (!ops.args || ops.args.length < 1 || ops.args.length > 2) {
-    console.log("usage: webpack-make config [section]");
-    process.exit(2);
-}
 
 var srcdir = process.env.SRCDIR;
 var makefile = ops.deps;
@@ -42,7 +38,7 @@ if (makefile) {
 }
 
 var cwd = process.cwd();
-var config_path = path.resolve(cwd, ops.args[0]);
+var config_path = path.resolve(cwd, ops.config);
 var config = require(config_path);
 
 // The latest input file time updated and used below