Watch mode makes fast iterations much faster. Compared to plain `make`,
watch mode and disabling eslint reduces e. g. tuned from 19s to 3s, or
systemd from 48s to 17s on the first rebuild, and 5s (!) on the second.