@martinpitt 2020-04-27
Resolved conflict and dropped erroneous whitespace change.