@martinpitt 2020-04-28
Issues from IRC:

 * Enable polling if `stat -fc %T .` == `nfs`
 * Don't exit on errors