@mvollmer 2020-04-28
>  * Enable polling if `stat -fc %T .` == `nfs`

This is not important anymore, since I will not be running `webpack-make -w` inside the VM.  Not using a VM is just fine and much faster.