@mvollmer 2020-04-28
>  * Don't exit on errors

I feel this is important.  If I have to baby-sit the watch process, it's just as distracting as waiting a minute for the results, maybe even more so bc I get so irritated... :-)