@marusak 2020-04-28
> Don't exit on errors

Absolutely. The 20s advantage disappears immediately when I don't realize that it failed and will wait for too long until it propagates :)
I liked the old '--watch' where one would just start it and hack on whatever pages and it would just rebuild it. And when things got broken, it would show big red error, but still would keep watching.