@mvollmer 2020-04-28
> And when things got broken, it would show big red error, but still would keep watching.

Except it would often break down and not actually recover from the error, at which point you need to restart it, IIRC. Hopefully it's better now.