@mvollmer 2020-04-28
I have pushed two commits, and with those I am happy I guess.

Fundamentally, I want incremental repacking but I don't care much about the file watching. I want to have a webpack process in the background for the page that I am working on, and I want to send it a signal of some sort every now and then when I want it to do its thing.

But if watching is the price to pay to get incremental builds, that's fine.