@martinpitt 2020-04-28
I reworked this again to incorporate @mvollmer's fixups (I reworked webpack-watch a bit, though), added a "build started" message, simplified the output format, and addressed the typo from above. PTAL?